+++ RN_SDK/src/main/java/com/example/rnsdk/Models/TransaferPointChildPageDataModel.java
@@ -2,13 +2,13 @@ package com.example.rnsdk.Models;
 
 public class TransaferPointChildPageDataModel{
     public String image;
-    public double opacity;
+    public float opacity;
     public boolean isClickable;
     public String linkType;
     public String internalLink;
     public String externalLink;
 
-    public TransaferPointChildPageDataModel(String image, double opacity, boolean isClickable, String linkType, String internalLink, String externalLink) {
+    public TransaferPointChildPageDataModel(String image, float opacity, boolean isClickable, String linkType, String internalLink, String externalLink) {
         this.image = image;
         this.opacity = opacity;
         this.isClickable = isClickable;
@@ -25,11 +25,11 @@ public class TransaferPointChildPageDataModel{
         this.image = image;
     }
 
-    public double getOpacity() {
+    public float getOpacity() {
         return opacity;
     }
 
-    public void setOpacity(double opacity) {
+    public void setOpacity(float opacity) {
         this.opacity = opacity;
     }
